Final refactor of categorical tests
