+++ pandas/tests/categorical/test_constructors.py
@@ -1,4 +1,5 @@
 # -*- coding: utf-8 -*-
+
 import pytest
 from datetime import datetime
 
@@ -507,3 +508,10 @@ class TestCategoricalConstructors(object):
         assert not cat.ordered
         cat = Categorical([0, 1, 2], ordered=True)
         assert cat.ordered
+
+    @pytest.mark.xfail(reason="Imaginary values not supported in Categorical")
+    def test_constructor_imaginary(self):
+        values = [1, 2, 3 + 1j]
+        c1 = Categorical(values)
+        tm.assert_index_equal(c1.categories, Index(values))
+        tm.assert_numpy_array_equal(np.array(c1), np.array(values))
